Approximate rewards UI properly compounds rewards
