+++ app/actions/ada/delegation-transaction-actions.js
@@ -1,16 +1,13 @@
 // @flow
 import { AsyncAction, Action } from '../lib/Action';
-
-export type PoolRequest =
-  void |
-  {| id: string |} |
-  Array<{|
-    id: string,
-    part: number,
-  |}>;
+import PublicDeriverWithCachedMeta from '../../domain/PublicDeriverWithCachedMeta';
+import type { PoolRequest } from '../../api/ada/lib/storage/bridge/delegationUtils';
 
 export default class DelegationTransactionActions {
-  createTransaction: AsyncAction<PoolRequest> = new AsyncAction();
+  createTransaction: AsyncAction<{|
+    publicDeriver: PublicDeriverWithCachedMeta,
+    poolRequest: PoolRequest,
+  |}> = new AsyncAction();
   signTransaction: AsyncAction<{| password: string |}> = new AsyncAction();
   complete: AsyncAction<void> = new AsyncAction();
   reset: Action<void> = new Action();
