+++ app/components/wallet/staking/dashboard/StakingDashboard.scss
@@ -8,6 +8,10 @@
   flex-direction: column;
 }
 
+.warningBox {
+  margin-bottom: 16px;
+}
+
 .contentWrap {
   max-width: 1178px;
   width: 100%;
